# Generate source to populate a NpgsqlDatabaseInfo