--- a/.github/workflows/update-catalogs.yml
+++ b/.github/workflows/update-catalogs.yml
@@ -5,7 +5,6 @@ on:
     branches:
       - main
       - 'hotfix/**'
-      - DAT_FILE_UPDATER
   schedule:
     - cron: "0 0 * * *"