--- a/src/Npgsql.SourceGenerators/PostgresMinimalDatabaseInfo.snbtxt
+++ b/src/Npgsql.SourceGenerators/PostgresMinimalDatabaseInfo.snbtxt
@@ -10,10 +10,14 @@ namespace Npgsql;
 
 partial class PostgresMinimalDatabaseInfo
 {
+    internal static readonly int CatalogVersion;
+
     static readonly PostgresType[] Types;
 
     static PostgresMinimalDatabaseInfo()
     {
+        CatalogVersion = {{ version }};
+
         ///////////////////////////
         // Referenced base types //
         ///////////////////////////