--- a/src/Npgsql/PostgresMinimalDatabaseInfo.cs
+++ b/src/Npgsql/PostgresMinimalDatabaseInfo.cs
@@ -1,11 +1,8 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using Npgsql.Internal;
 using Npgsql.PostgresTypes;
 using Npgsql.Util;
-using NpgsqlTypes;
 
 namespace Npgsql;
 
@@ -19,20 +16,11 @@ class PostgresMinimalDatabaseInfoFactory : INpgsqlDatabaseInfoFactory
         );
 }
 
-class PostgresMinimalDatabaseInfo : PostgresDatabaseInfo
+sealed partial class PostgresMinimalDatabaseInfo : PostgresDatabaseInfo
 {
-    static readonly PostgresBaseType[] Types = typeof(NpgsqlDbType).GetFields()
-        .Select(f => f.GetCustomAttribute<BuiltInPostgresType>())
-        .OfType<BuiltInPostgresType>()
-        .Select(a => new PostgresBaseType("pg_catalog", a.Name, a.OID))
-        .ToArray();
-
     protected override IEnumerable<PostgresType> GetTypes() => Types;
 
-    internal PostgresMinimalDatabaseInfo(NpgsqlConnector conn)
-        : base(conn)
-    {
-        HasIntegerDateTimes = !conn.PostgresParameters.TryGetValue("integer_datetimes", out var intDateTimes) ||
-                              intDateTimes == "on";
-    }
+    internal PostgresMinimalDatabaseInfo(NpgsqlConnector conn) : base(conn)
+        => HasIntegerDateTimes = !conn.PostgresParameters.TryGetValue("integer_datetimes", out var intDateTimes) ||
+                                 intDateTimes == "on";
 }