This adds a souce generator that parses the pg_type.dat and pg_range.dat
(copies from their src/include/catalog/ folder) files, that PostgreSQL
itself uses to initialize a cluster, to generate a NpgsqlDatabaseInfo.

Closes #3890