@Brar 2022-07-25
> I looked at this again with fresh eyes, and here's a bit of high-level thinking that you may not entirely like 😬

Well, if you make it appear like people only need to suggest things and you implement them, you may end up investing a lot of time in people's floating thoughts about the problem at hands and never get it solved in the end.
At some point you have to pull the ripcord, which is what I'm doing now, although I don't disagree with what you write.

There are definitely hundreds of ways to approach and solve this problem, one of them being to transfer information by hand (what we had before), one being to solve things 100% in C# (my initial approach), one being a hybrid C#/Perl model (my current approach) and another being a pure Perl approach.
All of them have advantages and disadvantages and none of them will ever be perfect.

For now, I'd say: "Let's get this into a state where it is committable and commit it."

If we run into actual problems that are measurable or visible in some other way, I'm happy to help solving them.

I'll go over your individual comments and see if/how I can address them.