@roji 2022-07-25
I'm just a bit frustrated that we had a design discussion about how to go about this, and it didn't occur to me to make these comments back then.

> If we run into actual problems that are measurable or visible in some other way, I'm happy to help solving them.

IMHO the problem is that as always with build perf, we won't notice the adverse effects; the inner loop will just become slower because we're doing lots of new JSON deserialization in the background. I don't know enough about source generators to fully understand how things work perf-wise (e.g. how often the source generator will get triggered, whether SHA1-based caching is a viable strategy), and my gut feeling is that I'd rather go into those tricky/risky questions only if I have to, for something that justifies it. But it's up to you.