@Brar 2022-07-25
> I'm just a bit frustrated that we had a design discussion about how to go about this, and it didn't occur to me to make these comments back then.

I don't have any hard feelings. Thoughts evolve. That's how it works.

> IMHO the problem is that as always with build perf, we won't notice the adverse effects; the inner loop will just become slower because we're doing lots of new JSON deserialization in the background. I don't know enough about source generators to fully understand how things work perf-wise (e.g. how often the source generator will get triggered, whether SHA1-based caching is a viable strategy), and my gut feeling is that I'd rather go into those tricky/risky questions only if I have to, for something that justifies it.

I agree. That's why I'm asking for a measurable effect.

Here's my attempt to measure the build-perf-problem:

| Method |    Mean |    Error |   StdDev | Ratio | RatioSD |
|------- |--------:|---------:|---------:|------:|--------:|
| Before | 1.708 s | 0.0338 s | 0.0299 s |  1.00 |    0.00 |
|  After | 1.687 s | 0.0334 s | 0.0357 s |  0.99 |    0.03 |

<details>
  <summary>Benchmark code:</summary>

````````csharp
using System.Diagnostics;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace BuildMeasurement;

public class BuildRunner
{
    readonly Process _beforeProcess;
    readonly Process _afterProcess;

    public BuildRunner()
    {
        _beforeProcess = CreateProcess("..\\Npgsql_main\\src\\Npgsql\\Npgsql.csproj");
        _afterProcess = CreateProcess("..\\npgsql\\src\\Npgsql\\Npgsql.csproj");
    }

    static Process CreateProcess(string projectDir)
    {
        var proc = new Process();
        proc.StartInfo.FileName = "dotnet";
        proc.StartInfo.ArgumentList.Add("build");
        proc.StartInfo.ArgumentList.Add(projectDir);
        proc.StartInfo.RedirectStandardOutput = true;
        return proc;
    }

    [Benchmark(Baseline = true)]
    public void Before()
    {
        _beforeProcess.Start();
        _beforeProcess.StandardOutput.ReadToEnd();
        _beforeProcess.WaitForExit();
    }

    [Benchmark]
    public void After()
    {
        _afterProcess.Start();
        _afterProcess.StandardOutput.ReadToEnd();
        _afterProcess.WaitForExit();
    }
}

public static class Program
{
    public static void Main(string[] args)
    {
        var summary = BenchmarkRunner.Run<BuildRunner>();
    }
}
````````

</details>

As you can see, build perf does not seem to be a measurable problem.

There may be other perf problems that might go as far as slowing down the whole IDE but this seems to be a general problem of v1 source generators, that is rather [dependent on the solution size](https://github.com/dotnet/runtime/issues/56702#issuecomment-895578440), than on the actual work, the generator is performing.

In any case I'll have a look at v2 (incremental) source generators (`IIncrementalGenerator`) since they seem to solve some problems, which might help us with the other generators as well.