@Brar 2022-07-25
> more with the inner loop when working inside the IDE

Yeah, that's the second kind of problem I mentioned, which - as I understand the problems leading to the design of `IIncrementalGenerator` - depends mostly on the overall solution size and would either affect all of our source generators or none of them.

That's why I'll look into `IIncrementalGenerator`. It may well serve us beyond this PR.