@roji 2022-09-11
Closing as per https://github.com/npgsql/npgsql/issues/3890#issuecomment-1242988454.